[FC-Palette#22] feat: ChatController 하나로 통합
